+++ core/src/com/unciv/logic/city/managers/CityExpansionManager.kt
@@ -89,10 +89,10 @@ class CityExpansionManager : IsPartOfGameInfoSerialization {
     }
 
     @Readonly
-    fun getGoldCostOfTile(tile: Tile): Int {
+    fun getGoldCostOfTile(tile: Tile, extraTiles: Int = 0): Int {
         val baseCost = 50
         val distanceFromCenter = tile.aerialDistanceTo(city.getCenterTile())
-        var cost = baseCost * (distanceFromCenter - 1) + tilesClaimed() * 5.0
+        var cost = baseCost * (distanceFromCenter - 1) + (tilesClaimed() + extraTiles) * 5.0
 
         cost *= city.civ.gameInfo.speed.goldCostModifier
 
@@ -183,8 +183,8 @@ class CityExpansionManager : IsPartOfGameInfoSerialization {
             tile.getCity()!!.expansion.relinquishOwnership(tile)
 
         if (tile.improvement == Constants.barbarianEncampment)
-            tile.setImprovementBasic(null)
-            
+            tile.removeImprovement()
+
         city.tiles = city.tiles.withItem(tile.position)
         tile.setOwningCity(city)
         city.population.autoAssignPopulation()