+++ core/src/com/unciv/ui/popups/AnimatedMenuPopup.kt
@@ -12,18 +12,18 @@ import com.badlogic.gdx.scenes.scene2d.Touchable
 import com.badlogic.gdx.scenes.scene2d.actions.Actions
 import com.badlogic.gdx.scenes.scene2d.ui.Container
 import com.badlogic.gdx.scenes.scene2d.ui.Table
+import com.badlogic.gdx.scenes.scene2d.utils.Layout
 import com.badlogic.gdx.scenes.scene2d.utils.NinePatchDrawable
 import com.badlogic.gdx.utils.Align
+import com.unciv.UncivGame
+import com.unciv.models.metadata.GameSettings.LongPressIndicatorSetting
 import com.unciv.ui.components.SmallButtonStyle
 import com.unciv.ui.components.UncivTooltip
 import com.unciv.ui.components.UncivTooltip.Companion.getEdgePoint
+import com.unciv.ui.components.extensions.allChildren
 import com.unciv.ui.components.extensions.setSize
 import com.unciv.ui.components.extensions.toTextButton
-import com.unciv.ui.components.input.KeyCharAndCode
-import com.unciv.ui.components.input.KeyboardBinding
-import com.unciv.ui.components.input.keyShortcuts
-import com.unciv.ui.components.input.onActivation
-import com.unciv.ui.components.input.onRightClick
+import com.unciv.ui.components.input.*
 import com.unciv.ui.images.ImageGetter
 import com.unciv.ui.screens.basescreen.BaseScreen
 import com.unciv.utils.Concurrency
@@ -43,27 +43,34 @@ import com.unciv.utils.Concurrency
  *  inverting the fade-and-scale-in. Callbacks registered with [Popup.closeListeners] will run before the animation starts.
  *  Use [afterCloseCallback] instead if you need a notification after the animation finishes and the Popup is cleaned up.
  *
- *  Using the new constructor taking a target [Actor], it will receive a decoration symbol over its bottom right,
+ *  Using the new `onRightClick` replacement [addContextMenu], the target actor will receive a decoration symbol over its bottom right,
  *  permanent on Android but on hover only on desktop.
  */
 open class AnimatedMenuPopup private constructor (stage: Stage) : Popup(stage, Scrollability.None) {
     companion object {
-        private const val indicatorAndroid = "OtherIcons/ContextMenuAvailableA"
+        const val indicatorAndroid = "OtherIcons/ContextMenuAvailableA"
         private const val indicatorDesktop = "OtherIcons/ContextMenuAvailableD"
         private const val maxIndicatorSize = 32f
-        private const val indicatorXOffset = 10f
+        // These are partially to take up the "slack" inside the texture
+        // (which is square, but the visual content for the Desktop one has a wide transparent margin)
+        private const val indicatorXOffsetRelativeAndroid = 0.15f
+        private const val indicatorXOffsetRelativeDesktop = 0.35f
         private val indicatorColor = Color.WHITE
-
-        /** Get stage coords of an [actor]'s top right corner, to help position an [AnimatedMenuPopup].
-         *  Note the Popup will center over this point.
-         */
-        fun getActorTopRight(actor: Actor): Vector2 = actor.getAlignedPosition(Align.topRight)
 
         private fun Actor.getAlignedPosition(align: Int) = localToStageCoordinates(getEdgePoint(align))
 
-        fun Group.addContextMenu(factory: () -> AnimatedMenuPopup) {
-            Helpers.addIndicator(this)
-            onRightClick { factory() }
+        fun Group.addContextMenu(indicatorMinSizeRelative: Float = 0.5f, factory: () -> AnimatedMenuPopup) {
+            Helpers.addIndicator(this, indicatorMinSizeRelative)
+            onRightClick {
+                if (isActive())
+                    factory()
+            }
+        }
+
+        fun Group.adjustContextMenuIndicators() {
+            for (icon in allChildren().filter { it.name == indicatorAndroid }) {
+                Helpers.adjustPosition(icon)
+            }
         }
     }
 
@@ -71,8 +78,8 @@ open class AnimatedMenuPopup private constructor (stage: Stage) : Popup(stage, S
      *  @param stage The stage this will be shown on, passed to Popup and used for clamping **`position`**
      *  @param position stage coordinates to show this centered over - clamped so that nothing is clipped outside the [stage]
      */
-    @Deprecated("Use the constructor taking a target actor instead.") // TODO remove deprecation and make ctor private
-    constructor(stage: Stage, position: Vector2) : this(stage) {
+    @Suppress("unused") // False positive
+    private constructor(stage: Stage, position: Vector2) : this(stage) {
         clickBehindToClose = true
         keyShortcuts.add(KeyCharAndCode.BACK) { close() }
         innerTable.remove()
@@ -86,8 +93,7 @@ open class AnimatedMenuPopup private constructor (stage: Stage) : Popup(stage, S
      *  @param stage The stage this will be shown on, passed to Popup and used for clamping **`position`**
      *  @param actor Target [Actor]. This popup will be shown centered relative to it, and it will get a visual indicator.
      */
-    @Suppress("deprecation")
-    constructor(stage: Stage, actor: Group, align: Int = Align.topRight) : this(stage, actor.getAlignedPosition(align))
+    constructor(stage: Stage, actor: Actor, align: Int = Align.topRight) : this(stage, actor.getAlignedPosition(align))
 
     private val container: Container<Table> = Container()
     private val animationDuration = 0.33f
@@ -134,29 +140,49 @@ open class AnimatedMenuPopup private constructor (stage: Stage) : Popup(stage, S
 
 
     private object Helpers {
-        fun addIndicator(actor: Group) =
-            if (Gdx.app.type == Application.ApplicationType.Android) addIndicatorAndroid(actor)
-            else addIndicatorDesktop(actor)
+        fun addIndicator(actor: Group, indicatorMinSizeRelative: Float) {
+            val setting = UncivGame.Current.settings.showLongPressIndicators
+            if (setting == LongPressIndicatorSetting.Off) return
+            if (setting == LongPressIndicatorSetting.Debug || Gdx.app.type == Application.ApplicationType.Android)
+                addIndicatorAndroid(actor, indicatorMinSizeRelative)
+            else addIndicatorDesktop(actor, indicatorMinSizeRelative)
+        }
 
-        private fun addIndicatorAndroid(actor: Group) {
-            val size = maxIndicatorSize.coerceAtMost(actor.height / 2)
+        fun adjustPosition(actor: Actor) {
+            val setting = UncivGame.Current.settings.showLongPressIndicators
+            if (setting == LongPressIndicatorSetting.Off) return
+            if (setting != LongPressIndicatorSetting.Debug && Gdx.app.type == Application.ApplicationType.Android) return
+            val x = actor.parent.width + actor.width * indicatorXOffsetRelativeAndroid
+            actor.setPosition(x , 0f, Align.bottomRight)
+        }
+
+        private fun getIndicatorSize(actor: Group, indicatorMinSizeRelative: Float): Float {
+            val actorHeight = (actor as? Layout)?.prefHeight ?: actor.height
+            return maxIndicatorSize.coerceAtMost(actorHeight * indicatorMinSizeRelative)
+        }
+
+        private fun addIndicatorAndroid(actor: Group, indicatorMinSizeRelative: Float) {
             val icon = ImageGetter.getImage(indicatorAndroid, indicatorColor)
+            icon.name = indicatorAndroid
+            val size = getIndicatorSize(actor, indicatorMinSizeRelative)
             icon.setSize(size)
-            icon.setPosition(actor.width + indicatorXOffset, 0f, Align.bottomRight)
+            val x = ((actor as? Layout)?.prefWidth ?: actor.width) + size * indicatorXOffsetRelativeAndroid
+            icon.setPosition(x , 0f, Align.bottomRight)
             actor.addActor(icon)
         }
 
-        private fun addIndicatorDesktop(actor: Group) {
-            val size = maxIndicatorSize.coerceAtMost(actor.height / 2)
+        private fun addIndicatorDesktop(actor: Group, indicatorMinSizeRelative: Float) {
             val icon = ImageGetter.getImage(indicatorDesktop, indicatorColor)
+            val size = getIndicatorSize(actor, indicatorMinSizeRelative)
             icon.setSize(size)
+            val offset = size * indicatorXOffsetRelativeDesktop
             actor.addListener(
                 UncivTooltip(
                     target = actor,
                     content = icon,
                     targetAlign = Align.bottomRight,
                     tipAlign = Align.bottomRight,
-                    offset = Vector2(indicatorXOffset, 0f)
+                    offset = Vector2(offset, 0f)
                 )
             )
         }